backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.0
sqlmodel==0.0.22
pydantic==2.10.1
pydantic-core==2.27.1
pydantic-settings==2.6.1
SQLAlchemy==2.0.36
openai==1.57.4
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.14.0
httpx==0.28.0
mcp>=1.0.0
python-dotenv==1.0.1
black==24.10.0
mypy==1.13.0
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==5.0.0
requests==2.32.3
python-jwt==4.0.0
cryptography==43.0.3
email-validator==2.2.0
google-generativeai>=0.8.6